updated arch linux install file

+++ arch_install.sh
@@ -1,22 +1,22 @@
 #!/bin/bash
 
-stage=$1
+if [[ -e current_stage ]]
+then
+    stage=$(cat current_stage)
+else
+    stage=0
+fi
+
 cur=/home/florian/.dotfiles
 home_florian=/home/florian
 
-# Basic install (encryption?)
-# add user florian (-m -g users -G -p password)
-# vim
-# sudo/visudo
-# nvidia
-# reboot
-
-# dns,ens33.network,systemd-networkd
-# xterm termite vim
-# xorg xorg-server xorg-xinit
-# gnome gnome-extra, gdm enable/start
-# ||
-# kde, kdm enable/start
+# STAGE 0
+if [[ $stage == 0 ]]
+then
+    echo "git clone https://github.com/arty-hlr/.dotfiles.git"
+    echo -n 1 > current_stage
+    exit
+fi
 
 # STAGE 1
 if [[ $stage == 1 ]]
@@ -26,6 +26,7 @@ then
     pacman --noconfirm -S i3 dmenu xclip
     mkdir -p ~/.config/i3
     cp $cur/i3_config ~/.config/i3/config
+    echo -n 2 > current_stage
     reboot now
 fi
 
@@ -33,56 +34,56 @@ fi
 if [[ $stage == 2 ]]
 then
     pacman --noconfirm -S net-tools apache youtube-dl wget transmission-cli transmission-gtk qbittorrent irssi hexchat imagemagick gimp vlc subdl subdownloader mate-terminal tmux ranger caja perl-rename git cmake gdb gparted htop libreoffice-still vim-latexsuite calibre knotes clamav bc sagemath typespeed mlocate
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     pacman --noconfirm -S yay --noconfirm
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     sudo -u florian yay --noconfirm -S discord skypeforlinux-stable-bin slack-desktop realvnc-vnc-viewer hyx zulucrypt etcher cherrytree gtypist tpgt gdb-multiarch arm-linux-gnueabi-gcc aarch64-linux-gnu-gcc 
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     mhwd-kernel -i linux50-rt
-    echo "Installed kernels:"
-    mhwd-kernel -li
-    echo "Reboot now"
+    echo -n 3 > current_stage
+    reboot now
 fi
 
 # STAGE 3
 if [[ $stage == 3 ]]
 then
-    echo "Did you download vmware? (y/n)"
-    read res
-    if [[ $res = n ]]
-    then
-        echo "Download vmware-installer from https://www.vmware.com/products/workstation-pro/workstation-pro-evaluation.html"
-        exit
-    fi
+    echo "Download vmware-installer from https://www.vmware.com/products/workstation-pro/workstation-pro-evaluation.html"
+    echo -n 4 > current_stage
+    exit
+fi
+
+# STAGE 4
+if [[ $stage == 4 ]]
+then
     # vmware
     chmod +x $cur/*.bundle
     $cur/*.bundle
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     # pwntools
     pacman --noconfirm -S python2-pip
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     pip2 install pwntools
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     pacman --noconfirm -S python-pip
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     pip3 install --upgrade git+https://github.com/arthaud/python3-pwntools.git
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     $cur/binutils.sh arm
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     rm /tmp/binutils-build/*/config.cache
     $cur/binutils.sh aarch64
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     rm /tmp/binutils-build/*/config.cache
     $cur/binutils.sh mips
     rm /tmp/binutils-build/*/config.cache
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     # .vimrc
     cp $cur/.vimrc $home_florian/.vimrc
     git clone https://github.com/VundleVim/Vundle.vim.git $home_florian/.vim/bundle/Vundle.vim
     vim +PluginInstall +qall
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     $home_florian/.vim/bundle/YouCompleteMe/install.py --clangd-completer 
-    echo "Next?"; read ok
+    # echo "Next?"; read ok
     cp $cur/.ycm_extra_conf.py $home_florian/.vim/.ycm_extra_conf.py
     # .tmux.conf
     git clone https://github.com/tmux-plugins/tpm $home_florian/.tmux/plugins/tpm
@@ -90,11 +91,12 @@ then
     cp $cur/.tmux.conf $home_florian/.tmux.conf
     # athame
     sudo -u florian yay --noconfirm -S readline-athame-git
-    echo "Reboot now"
+    echo -n 5 > current_stage
+    reboot now
 fi
 
-# STAGE 4
-if [[ $stage == 4 ]]
+# STAGE 5
+if [[ $stage == 5 ]]
 then
     # .bashrc
     rm $home_florian/.bashrc
